camera-system/infra/requirements.txt
# AWS CDK dependencies
aws-cdk-lib>=2.120.0
constructs>=10.0.0
